+++ x/bandwidth/model/bandwidth.go
@@ -14,7 +14,7 @@ type BandwidthInfo struct {
 
 // BlockInfo - stores info about number of tx in last block
 type BlockInfo struct {
-	TotalMsgSignedByApp  uint32         `json:"total_tx_signed_by_app"`
-	TotalMsgSignedByUser uint32         `json:"total_tx_signed_by_user"`
+	TotalMsgSignedByApp  int64          `json:"total_tx_signed_by_app"`
+	TotalMsgSignedByUser int64          `json:"total_tx_signed_by_user"`
 	CurMsgFee            linotypes.Coin `json:"cur_msg_fee"`
 }
